backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
langchain==0.2.11
langchain-core==0.2.23
httpx==0.27.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==24.1.0